Gallery: Open external uris directly
In order to achieve some kind of metadata to be shown for the media if you opened it using an external source (from files app for example), it involved an extra step to parse down the uri to a direct file path. This worked quite well for most cases, but there's some cases where (for ex: content://package_name/custom_file_path/filename.ext) the file location could not be parsed, which resulted in a crash.

Because of these cases I've decided to not show any kind of metadata from external uris.

Fixes #79

Signed-off-by: IacobIonut01 <paulionut2003@gmail.com>